requirements.txt
PyQt6~=6.7.1
python-dotenv~=1.0.1
pandas~=2.2.3
plotly~=5.24.1
win10toast~=0.9